+++ app/src/main/java/com/github/mobile/ui/user/HomePagerAdapter.java
@@ -25,6 +25,7 @@ import android.view.ViewGroup;
 import com.github.mobile.R;
 import com.github.mobile.ui.FragmentPagerAdapter;
 import com.github.mobile.ui.repo.RepositoryListFragment;
+import com.github.mobile.ui.repo.StarredRepositoryListFragment;
 
 import java.util.HashSet;
 import java.util.Set;
@@ -64,9 +65,11 @@ public class HomePagerAdapter extends FragmentPagerAdapter {
         case 1:
             return new RepositoryListFragment();
         case 2:
-            return defaultUser ? new MyFollowersFragment()
+            return defaultUser ? new StarredRepositoryListFragment()
                     : new MembersFragment();
         case 3:
+            return new MyFollowersFragment();
+        case 4:
             return new MyFollowingFragment();
         default:
             return null;
@@ -112,7 +115,7 @@ public class HomePagerAdapter extends FragmentPagerAdapter {
 
     @Override
     public int getCount() {
-        return defaultUser ? 4 : 3;
+        return defaultUser ? 5 : 3;
     }
 
     @Override
@@ -123,9 +126,11 @@ public class HomePagerAdapter extends FragmentPagerAdapter {
         case 1:
             return resources.getString(R.string.tab_repositories);
         case 2:
-            return resources.getString(defaultUser ? R.string.tab_followers_self
+            return resources.getString(defaultUser ? R.string.tab_stars
                     : R.string.tab_members);
         case 3:
+            return resources.getString(R.string.tab_followers_self);
+        case 4:
             return resources.getString(R.string.tab_following_self);
         default:
             return null;